registration function in util
