+++ torch_points3d/utils/geometry.py
@@ -16,3 +16,32 @@ def euler_angles_to_rotation_matrix(theta):
 
     R = torch.mm(R_z, torch.mm(R_y, R_x))
     return R
+
+
+def get_cross_product_matrix(k):
+    return torch.tensor([[0, -k[2], k[1]], [k[2], 0, -k[0]], [-k[1], k[0], 0]], device=k.device)
+
+
+def rodrigues(axis, theta):
+    """
+    given an axis of norm one and an angle, compute the rotation matrix using rodrigues formula
+    source : https://en.wikipedia.org/wiki/Rodrigues%27_rotation_formula
+    """
+    K = get_cross_product_matrix(axis)
+    t = torch.tensor([theta], device=axis.device)
+    R = torch.eye(3, device=axis.device) + torch.sin(t) * K + (1 - torch.cos(t)) * K.mm(K)
+    return R
+
+
+def get_trans(x):
+    """
+    get the rotation matrix from the vector representation using the rodrigues formula
+    """
+    T = torch.eye(4, device=x.device)
+    T[:3, 3] = x[3:]
+    axis = x[:3]
+    theta = torch.norm(axis)
+    if theta > 0:
+        axis = axis / theta
+    T[:3, :3] = rodrigues(axis, theta)
+    return T
